pyproject.toml
[project]
name = "django-whiteneuron"
version = "0.1.0"
description = "django-whiteneuron là một gói mở rộng giúp nâng cấp giao diện và chức năng Django Admin, mang đến trải nghiệm quản trị hiện đại, trực quan và tối ưu hiệu suất. 🚀"
authors = [
    {name = "White Neuron", email = "anhnt@whiteneuron.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
keywords = ["django", "plugin", "admin", "dashboard", "whiteneuron"]
dependencies = [
    "django>=5.0,<6.0",
    "django-celery-beat>=2.7.0",
    "django-guardian>=2.4.0",
    "django-imagekit>=5.0.0",
    "django-import-export>=4.3.3",
    "django-modeltranslation>=0.19.11",
    "django-ninja>=1.3.0",
    "django-simple-history>=3.7.0",
    "django-unfold>=0.49.1",
    "gunicorn>=23.0.0",
    "hiredis>=3.1.0",
    "markdown>=3.7",
    "openai>=1.58.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "redis>=5.2.1",
]
requires-python = ">=3.9"

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[project.urls]
homepage = "https://github.com/tanhtm/django-whiteneuron"
repository = "https://github.com/tanhtm/django-whiteneuron"

[tool.setuptools]
packages = ["whiteneuron"]
include-package-data = true

[tool.setuptools.package-data]
"whiteneuron" = ["templates/**/*", "static/**/*"]

[project.optional-dependencies]
dev = ["pytest", "black", "flake8", "pre-commit"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
